Comment service fixes with clients

+++ comment-service/src/main/java/com/comments/entity/Content.java
@@ -9,6 +9,6 @@ import java.io.Serializable;
 @Data
 public class Content implements Serializable {
 
-    private int contentId;
+    private Long contentId;
     private ContentType contentType;
 }
